Create CODE
The projct is targeted towards teenagers who enjoy playing car games and 
want a new twist on the  game genre. I kept this in mind when designing the user interface, drawing the graphics and 
creating the physics. Another audience that could have interest in the project is 
someone who is beginning to look at game , particularly racing/simulator games and needs 
guidance on how to start; then 
implement more advanced features such as . I
 would consider myself as a beneficiary as I have improved 
a lot of my existing programming skills and gained a lot of additional knowledge of various topics 
in Object Orientated Programming.
The purpose of this project was the creation of an easy to use, 
pick up and play game that could be played by all ages as long as they have an digital screen(
 android phone and a touch screen). The reason was as stated above there are more casual gamers than
 traditional gamers meaning a larger potential market. Developing on the Android platform will mean we have to program in JAVA or C# 
and get to grips with the Android SDK. I will also have to learn the fundamentals of game programming and stick to the lifecycle 
associated with developing games. The Android SDK does not provide a comprehensive set of tools to create a game straight away; it
 does require 
a lot of extra work. I could potentially create my own framework or perhaps use an existing one, to ease the process and produce 
software that is appealing and efficient
In our project we have and object oriented programming language which is c++. c++ is general purpose programming language. 
It can be used to develop operating system, browsers games, and so on.
 C++ also supports different ways of programming like procedural, functional, etc. 
since c++ is high level language that will teach you the basics of object oriented programming. It's a good idea to learn it.

We have also studied about different type of header files and their feature and uses such as:"iostream" "conio" "dos" "windows" "time"